refactor : 로또 생성 방식을 인터페이스로 분리
