+++ src/main/java/domain/generator/LottoGenerator.java
@@ -1,4 +1,4 @@
-package service;
+package domain.generator;
 
 import java.util.Collections;
 import java.util.List;
@@ -6,10 +6,8 @@ import java.util.stream.Collectors;
 
 import domain.Lotto;
 import domain.LottoNumber;
-import domain.Lottos;
-import domain.OrderForm;
 
-public interface LottoStrategy {
+public interface LottoGenerator {
 	List<LottoNumber> LOTTO_BUCKET = LottoNumber.ofList();
 
 	default Lotto createAutoLotto(int count) {
@@ -18,8 +16,4 @@ public interface LottoStrategy {
 			.limit(count)
 			.collect(Collectors.toUnmodifiableList()));
 	}
-
-	public Lottos createLottos(Lottos manualLottos, OrderForm orderForm);
-
-	public Lottos createLottos(int count, int size);
 }
