+++ src/main/java/service/LottoService.java
@@ -1,36 +1,22 @@
 package service;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import domain.Lotto;
 import domain.LottoResult;
 import domain.Lottos;
 import domain.OrderForm;
 
 import domain.WinningLotto;
+import domain.generator.AutoLottoGenerator;
+import domain.generator.ManualLottoGenerator;
 
-public class LottoService implements LottoStrategy {
+public class LottoService {
 
 	public Lottos createLottos(Lottos manualLottos, OrderForm orderForm) {
-		return createManualLottos(manualLottos, createLottos(orderForm.calculateAutoLottoCount(), Lotto.SIZE));
+		Lottos autoLottos = new AutoLottoGenerator().createLottos(orderForm.calculateAutoLottoCount(), Lotto.SIZE);
+		return new ManualLottoGenerator().createManualLottos(manualLottos, autoLottos);
 	}
 
-	public Lottos createLottos(int count, int size) {
-		List<Lotto> lottos = new ArrayList<>();
-		for (int i = 0; i < count; i++) {
-			lottos.add(createAutoLotto(size));
-		}
-		return new Lottos(lottos);
-	}
-
-		private Lottos createManualLottos(Lottos manualLottos, Lottos autoLottos) {
-			List<Lotto> lottos = new ArrayList<>(manualLottos.getLottos());
-			lottos.addAll(autoLottos.getLottos());
-			return new Lottos(lottos);
-		}
-
-		public LottoResult createLottoResult(Lottos lottos, WinningLotto winningLotto) {
+	public LottoResult createLottoResult(Lottos lottos, WinningLotto winningLotto) {
 		return LottoResult.from(lottos.countRank(winningLotto));
 	}
 }
