+++ src/test/java/domain/LottoServiceTest.java
@@ -5,16 +5,19 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import domain.generator.AutoLottoGenerator;
+import domain.generator.LottoGenerator;
+import domain.generator.ManualLottoGenerator;
 import service.LottoService;
 
 public class LottoServiceTest {
 
-	private final LottoService lottoService = new LottoService();
+	private final AutoLottoGenerator lottoGenerator = new AutoLottoGenerator();
 
 	@DisplayName("로또를 발행한다")
 	@Test
 	void createLotto() {
-		assertThat(lottoService.createLottos(5, Lotto.SIZE))
+		assertThat(lottoGenerator.createLottos(5, Lotto.SIZE))
 			.isInstanceOf(Lottos.class);
 	}
 
@@ -22,10 +25,11 @@ public class LottoServiceTest {
 	@Test
 	void createManualAndAutoLottos() {
 		//given
-		Lottos manualLottos = lottoService.createLottos(4, Lotto.SIZE);
+		Lottos manualLottos = lottoGenerator.createLottos(4, Lotto.SIZE);
 		//when
 		OrderForm orderForm = new OrderForm(new Payment(5000), 2);
-		Lottos manualAndAutoMixLottos = lottoService.createLottos(manualLottos, orderForm);
+		Lottos autoLottos = new AutoLottoGenerator().createLottos(orderForm.calculateAutoLottoCount(), Lotto.SIZE);
+		Lottos manualAndAutoMixLottos =  new ManualLottoGenerator().createManualLottos(manualLottos, autoLottos);
 		//then
 		assertThat(manualAndAutoMixLottos.getSize()).isEqualTo(7);
 	}
